Move runtime compatibility patches into dedicated module
