+++ straxen/__init__.py
@@ -1,32 +1,9 @@
 # mypy: disable-error-code="no-redef"
 __version__ = "3.2.7"
 
-import numpy as np
-import pandas as pd
-import strax
-import strax.utils as strax_utils
-
-# NumPy >= 2.4 removed np.in1d; keep an alias for dependencies (e.g. strax)
-# that still call it.
-if not hasattr(np, "in1d"):
-    np.in1d = np.isin
-
-# pandas >= 3 may return ExtensionArray (e.g. ArrowStringArray) for `.values`.
-# Make strax.to_str_tuple accept those arrays.
-_orig_to_str_tuple = strax_utils.to_str_tuple
-
-
-def _to_str_tuple_compat(x):
-    if isinstance(x, pd.api.extensions.ExtensionArray):
-        return tuple(x.tolist())
-    return _orig_to_str_tuple(x)
-
-
-strax_utils.to_str_tuple = _to_str_tuple_compat
-strax.to_str_tuple = _to_str_tuple_compat
-
 from utilix import uconfig
 from .common import *
+from ._compat import apply_runtime_compat
 
 from .itp_map import *
 from .matplotlib_utils import *
@@ -71,3 +48,5 @@ from .entry_points import load_entry_points
 
 load_entry_points()
 del load_entry_points
+
+apply_runtime_compat()
